Remove some redundancy
