+++ lib/transpiler.js
@@ -1,26 +1,15 @@
 'use strict';
 
 const babel = require('gulp-babel');
-const replace = require('gulp-replace');
 const rename = require('gulp-rename');
 const streamCombiner = require('./stream-combiner');
-const sourcemaps = require('gulp-sourcemaps');
 const path = require('path');
+const sourcemaps = require('./sourcemaps');
 
 
 const BABEL_OPTS = {
   configFile: path.resolve(__dirname, '..', '.babelrc'),
 };
-
-const SOURCEMAP_OPTS = {
-  sourceRoot: function (file) { // eslint-disable-line object-shorthand
-    // Point to source root relative to the transpiled file
-    return path.relative(path.join(file.cwd, file.path), file.base);
-  },
-  includeContent: true,
-};
-
-const HEADER = 'require(\'source-map-support\').install();\n\n';
 
 const renameEsX = function () {
   return rename(function (path) {
@@ -28,19 +17,17 @@ const renameEsX = function () {
   });
 };
 
-module.exports = function (opts) {
-  opts = opts || {};
+module.exports = function (opts = {}) {
+  const {sourceMapInit, sourceMapHeader, sourceMapWrite} = sourcemaps(opts.sourceMapOpts);
 
-  const babelOpts = Object.assign({}, BABEL_OPTS, opts.babelOpts);
-  const sourceMapOpts = Object.assign({}, SOURCEMAP_OPTS, opts.sourceMapOpts);
-  this.stream = function () {
-    return streamCombiner(function (source) {
+  this.stream = () => {
+    return streamCombiner((source) => {
       return source
-        .pipe(sourcemaps.init())
-        .pipe(babel(babelOpts))
-        .pipe(replace(/$/, HEADER))
+        .pipe(sourceMapInit)
+        .pipe(babel(Object.assign({}, BABEL_OPTS, opts.babelOpts)))
+        .pipe(sourceMapHeader)
         .pipe(renameEsX())
-        .pipe(sourcemaps.write(sourceMapOpts));
-    }.bind(this));
-  }.bind(this);
+        .pipe(sourceMapWrite);
+    });
+  };
 };
